+++ package.json
@@ -59,10 +59,10 @@
     "angular-split": "^4.0.0",
     "animate.css": "^3.6.1",
     "bootstrap": "^4.3.1",
-    "buffer": "^6.0.3",
+    "buffer": "^4.9.2",
     "codemirror": "^5.42.0",
     "core-js": "^2.5.4",
-    "crypto-js": "^4.1.1",
+    "crypto-js": "^4.0.0",
     "font-awesome": "4.7.0",
     "guacamole-common-js": "1.1.0",
     "jquery": "^3.4.1",
